Undo dir rename
