+++ go/vt/orchestrator/inst/audit_dao.go
@@ -24,10 +24,10 @@ import (
 
 	"github.com/rcrowley/go-metrics"
 
-	"vitess.io/vitess/go/vt/external/golib/sqlutils"
 	"vitess.io/vitess/go/vt/orchestrator/config"
 	"vitess.io/vitess/go/vt/orchestrator/db"
 	"vitess.io/vitess/go/vt/orchestrator/external/golib/log"
+	"vitess.io/vitess/go/vt/orchestrator/external/golib/sqlutils"
 )
 
 // syslogWriter is optional, and defaults to nil (disabled)
