+++ go/vt/orchestrator/inst/instance_topology_dao.go
@@ -23,10 +23,10 @@ import (
 	"strings"
 	"time"
 
-	"vitess.io/vitess/go/vt/external/golib/sqlutils"
 	"vitess.io/vitess/go/vt/orchestrator/config"
 	"vitess.io/vitess/go/vt/orchestrator/db"
 	"vitess.io/vitess/go/vt/orchestrator/external/golib/log"
+	"vitess.io/vitess/go/vt/orchestrator/external/golib/sqlutils"
 )
 
 // Max concurrency for bulk topology operations
