+++ go/vt/orchestrator/inst/pool_dao.go
@@ -19,10 +19,10 @@ package inst
 import (
 	"fmt"
 
-	"vitess.io/vitess/go/vt/external/golib/sqlutils"
 	"vitess.io/vitess/go/vt/orchestrator/config"
 	"vitess.io/vitess/go/vt/orchestrator/db"
 	"vitess.io/vitess/go/vt/orchestrator/external/golib/log"
+	"vitess.io/vitess/go/vt/orchestrator/external/golib/sqlutils"
 )
 
 // writePoolInstances will write (and override) a single cluster name mapping
