+++ go/vt/orchestrator/inst/resolve_dao.go
@@ -19,10 +19,10 @@ package inst
 import (
 	"github.com/rcrowley/go-metrics"
 
-	"vitess.io/vitess/go/vt/external/golib/sqlutils"
 	"vitess.io/vitess/go/vt/orchestrator/config"
 	"vitess.io/vitess/go/vt/orchestrator/db"
 	"vitess.io/vitess/go/vt/orchestrator/external/golib/log"
+	"vitess.io/vitess/go/vt/orchestrator/external/golib/sqlutils"
 )
 
 var writeResolvedHostnameCounter = metrics.NewCounter()
